+++ packages/xrpl/HISTORY.md
@@ -6,9 +6,11 @@ Subscribe to [the **xrpl-announce** mailing list](https://groups.google.com/g/xr
 
 ### Added
 * Adds utility function `convertTxFlagsToNumber`
+* New util `areAmountsEqual` to check if 2 amounts are strictly equal
 
 ### Changed
 * Deprecated `setTransactionFlagsToNumber`. Start using convertTxFlagsToNumber instead
+* `autofill` function in client not validating amounts correctly
 
 ## 4.1.0 (2024-12-23)